Add log in / out
Services and navbar controller are fully tested.

The e2e tests are missing. I don’t see any meaningful way to mock the firebase
websocket connection. Instead of mocking it, I will use a second database
that the e2e tests can reset at the end of a scenario.

The only thing mocked will be the calls to the verifier API and
authentication.

I will tackle the e2e test after the initial work on the rules.